+++ src/option.h
@@ -60,6 +60,7 @@ typedef enum {
    OPT_PSL_COMMENTS,
    OPT_NO_COLLAPSE,
    OPT_COVER_VERBOSE,
+   OPT_COVER_TIMESTAMP,
 
    OPT_LAST_NAME
 } opt_name_t;